+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - fix: spelling error for 'nodeselectChange', closes: #3736;
 - fix: update node icon from show false to show true;
+- fix: afterrender should be emitted when the layout is not configured;
 - perf: update related edges while drag-combo;
 - feat: combo supports collapsedSubstituteIcon showing after collapsed;
 - feat: remove animations while first rendering with (collapsed)combos;